Decoupling/ML/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host https

aiomqtt==2.1.0
    # via -r requirements.in
async-timeout==4.0.3
    # via redis
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dependency-injector==4.41.0
    # via -r requirements.in
fonttools==4.51.0
    # via matplotlib
importlib-resources==6.4.0
    # via matplotlib
joblib==1.4.2
    # via
    #   pyod
    #   scikit-learn
kiwisolver==1.4.5
    # via matplotlib
llvmlite==0.42.0
    # via numba
matplotlib==3.9.0
    # via
    #   pyod
    #   sranodec
numba==0.59.1
    # via pyod
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   numba
    #   outlier-utils
    #   pandas
    #   patsy
    #   pyarrow
    #   pyod
    #   scikit-learn
    #   scipy
    #   sranodec
    #   statsmodels
orjson==3.10.3
    # via -r requirements.in
outlier-utils==0.0.5
    # via -r requirements.in
packaging==24.0
    # via
    #   matplotlib
    #   statsmodels
paho-mqtt==2.1.0
    # via aiomqtt
pandas==2.2.2
    # via
    #   -r requirements.in
    #   statsmodels
patsy==0.5.6
    # via statsmodels
pillow==10.3.0
    # via matplotlib
pyarrow==16.1.0
    # via -r requirements.in
pyod==1.1.3
    # via -r requirements.in
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
redis==5.0.4
    # via -r requirements.in
scikit-learn==1.5.0
    # via pyod
scipy==1.13.1
    # via
    #   outlier-utils
    #   pyod
    #   scikit-learn
    #   sranodec
    #   statsmodels
six==1.16.0
    # via
    #   dependency-injector
    #   patsy
    #   pyod
    #   python-dateutil
sranodec==0.0.1
    # via -r requirements.in
statsmodels==0.14.2
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
typing-extensions==4.12.0
    # via aiomqtt
tzdata==2024.1
    # via pandas
zipp==3.18.2
    # via importlib-resources
